refactor!: More refactoring and cleanup.
