+++ src/api.js
@@ -1,0 +1,32 @@
+// export internal class and method definitions
+export { BitSet } from './table/BitSet.js';
+export { Table } from './table/Table.js';
+export { ColumnTable } from './table/ColumnTable.js';
+export { default as Reducer } from './verbs/reduce/reducer.js';
+export { default as parse } from './expression/parse.js';
+export { default as walk_ast } from './expression/ast/walk.js';
+
+// public API
+export { seed } from './util/random.js';
+export { default as fromArrow } from './arrow/from-arrow.js';
+export { default as fromCSV } from './format/from-csv.js';
+export { default as fromFixed } from './format/from-fixed.js';
+export { default as fromJSON } from './format/from-json.js';
+export { default as toArrow } from './arrow/to-arrow.js';
+export { default as toArrowIPC } from './arrow/to-arrow-ipc.js';
+export { default as toCSV } from './format/to-csv.js';
+export { default as toHTML } from './format/to-html.js';
+export { default as toJSON } from './format/to-json.js';
+export { default as toMarkdown } from './format/to-markdown.js';
+export { default as bin } from './helpers/bin.js';
+export { default as escape } from './helpers/escape.js';
+export { default as desc } from './helpers/desc.js';
+export { default as field } from './helpers/field.js';
+export { default as frac } from './helpers/frac.js';
+export { default as names } from './helpers/names.js';
+export { default as rolling } from './helpers/rolling.js';
+export { all, endswith, matches, not, range, startswith } from './helpers/selection.js';
+export { default as agg } from './verbs/helpers/agg.js';
+export { default as op } from './op/op-api.js';
+export { addAggregateFunction, addFunction, addWindowFunction } from './op/register.js';
+export { table, from } from './table/index.js';
