+++ src/index-browser.js
@@ -1,0 +1,2 @@
+export * from './api.js';
+export { load, loadArrow, loadCSV, loadFixed, loadJSON } from './format/load-url.js';
