--- src/index-node.js
@@ -1,8 +1,0 @@
-export * from './index.js';
-export {
-  load,
-  loadArrow,
-  loadCSV,
-  loadFixed,
-  loadJSON
-} from './format/load-file.js';
