+++ src/verbs/intersect.js
@@ -1,7 +1,8 @@
 import { dedupe } from './dedupe.js';
 import { semijoin } from './join-filter.js';
 
-export function intersect(table, others) {
+export function intersect(table, ...others) {
+  others = others.flat();
   const names = table.columnNames();
   return others.length
     ? dedupe(others.reduce((a, b) => semijoin(a, b.select(names)), table))
