+++ src/verbs/rename.js
@@ -1,8 +1,8 @@
 import { _select } from './select.js';
 import resolve from '../helpers/selection.js';
 
-export function rename(table, columns) {
+export function rename(table, ...columns) {
   const map = new Map();
   table.columnNames(x => (map.set(x, x), 0));
-  return _select(table, resolve(table, columns, map));
+  return _select(table, resolve(table, columns.flat(), map));
 }
