+++ src/verbs/select.js
@@ -3,8 +3,8 @@ import { columnSet } from '../table/ColumnSet.js';
 import error from '../util/error.js';
 import isString from '../util/is-string.js';
 
-export function select(table, columns) {
-  return _select(table, resolve(table, columns));
+export function select(table, ...columns) {
+  return _select(table, resolve(table, columns.flat()));
 }
 
 export function _select(table, columns) {
