+++ src/verbs/union.js
@@ -1,5 +1,6 @@
+import { concat } from './concat.js';
 import { dedupe } from './dedupe.js';
 
-export function union(table, others) {
-  return dedupe(table.concat(others));
+export function union(table, ...others) {
+  return dedupe(concat(table, others.flat()));
 }
