+++ test/table/bitset-test.js
@@ -1,5 +1,5 @@
 import assert from 'node:assert';
-import { BitSet } from '../../src/table/BitSet.js';
+import { BitSet } from '../../src/index.js';
 
 describe('BitSet', () => {
   it('manages a set of bits', () => {
