+++ test/verbs/groupby-test.js
@@ -1,7 +1,5 @@
 import assert from 'node:assert';
-import fromArrow from '../../src/arrow/from-arrow.js';
-import toArrow from '../../src/arrow/to-arrow.js';
-import { desc, op, table } from '../../src/index.js';
+import { desc, fromArrow, op, table, toArrow } from '../../src/index.js';
 
 describe('groupby', () => {
   it('computes groups based on field names', () => {
